% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getGenes}
\alias{getGenes}
\alias{getGenes,NGCHM_inferCNV-method}
\title{Get the gene names}
\usage{
getGenes(obj)

\S4method{getGenes}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
Gene Locations
}
\description{
Get the gene names
}
